AnalogInput value type changed to double
for IPX800 v5 support

+++ IPX800cs/IPX800Base.cs
@@ -36,7 +36,7 @@ public abstract class IPX800Base : IIPX800, IDisposable
         return ResponseParserFactory.CreateGetInputsParser(Protocol, inputType).ParseResponse(response);
     }
 
-    public int GetAnalogInput(AnalogInput input)
+    public double GetAnalogInput(AnalogInput input)
     {
         var command = CommandFactory.CreateGetAnalogInputCommand(input);
         var response = CommandSender.ExecuteCommand(command);
